@@ -3,8 +3,9 @@ module.exports = {
   env: {
     node: true
   },
-  'extends': [
-    'plugin:vue/essential',
+  extends: [
+    'plugin:vue/vue3-essential',
+    'eslint:recommended',
     '@vue/standard'
   ],
   rules: {
@@ -12,6 +13,6 @@ module.exports = {
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off'
   },
   parserOptions: {
-    parser: 'babel-eslint'
+    ecmaVersion: 'latest'
   }
 }
